refactor(app): change naming & files structure (NO-JIRA)
* refactor(auth): change auth-store, add path for actions(NO-JIRA)

* refactor(auth): remove eslint ignore rule(NO-JIRA)

* refactor(auth): change auth-guard(NO-JIRA)

* refactor(auth): change path for enum store feature names(NO-JIRA)

* refactor(auth): change name for utils(NO-JIRA)

* refactor(auth): change name for models(NO-JIRA)

* refactor(shared): change name for constants(NO-JIRA)

* refactor(shared): change name for utils(NO-JIRA)

* refactor(shared): change name for enums(NO-JIRA)

* refactor(shared): change name for validatos(NO-JIRA)

* refactor(scripts): add and applie new eslint rule(NO-JIRA)
