+++ src/app/auth/auth-store/auth.effects.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
 import { inject, Injectable } from '@angular/core'
 import { Router } from '@angular/router'
 import { Actions, createEffect, ofType } from '@ngrx/effects'
@@ -6,14 +5,14 @@ import { TuiAlertService } from '@taiga-ui/core'
 import { catchError, of } from 'rxjs'
 import { map, switchMap, tap } from 'rxjs/operators'
 
-import { alertsAuth } from '../dictionary/auth-alert.dictionary'
-import { authInitApiActions } from './auth-init-api.actions'
-import { authInitActions } from './auth-init.actions'
-import { logoutActions } from './logout.actions'
-import { signInApiActions } from './sign-in-api.actions'
-import { signInPageActions } from './sign-in-page.actions'
-import { signUpApiActions } from './sign-up-api.actions'
-import { signUpPageActions } from './sign-up-page.actions'
+import { alertsAuth } from '../utils/auth-alert.util'
+import { authInitApiActions } from './actions/auth-init-api.actions'
+import { authInitActions } from './actions/auth-init.actions'
+import { logoutActions } from './actions/logout.actions'
+import { signInApiActions } from './actions/sign-in-api.actions'
+import { signInPageActions } from './actions/sign-in-page.actions'
+import { signUpApiActions } from './actions/sign-up-api.actions'
+import { signUpPageActions } from './actions/sign-up-page.actions'
 import { CommercetoolsService } from 'src/app/core/commercetools/services/commercetools.service'
 
 @Injectable()
