+++ src/app/auth/auth-store/auth.selectors.ts
@@ -1,7 +1,7 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store'
 
-import type { AuthState } from '../model/auth-state.model'
-import { StoreFeatureNames } from 'src/app/shared/enum/store-feature-names.enum'
+import type { AuthState } from '../models/auth-state.model'
+import { StoreFeatureNames } from 'src/app/shared/enums/store-feature-names.enum'
 
 const selectAuthFeature = createFeatureSelector<AuthState>(StoreFeatureNames.Auth)
 
@@ -10,3 +10,5 @@ export const selectError = createSelector(selectAuthFeature, ({ error }: AuthSta
 export const selectIsLoggedIn = createSelector(selectAuthFeature, ({ isLoggedIn }: AuthState) => isLoggedIn)
 
 export const selectIsLoading = createSelector(selectAuthFeature, ({ isLoading }: AuthState) => isLoading)
+
+export const selectAuthData = createSelector(selectAuthFeature, ({ authData }: AuthState) => authData)
