+++ src/app/auth/components/sign-in-form/sign-in-form.component.ts
@@ -8,9 +8,9 @@ import { hasOneLowerCaseCharacter } from '../../../shared/validators/has-one-low
 import { hasOneUpperCaseCharacter } from '../../../shared/validators/has-one-uppercase-character.validator'
 import { minCharacterValidator } from '../../../shared/validators/min-character.validator'
 import { AuthFacade } from '../../auth-store/auth.facade'
-import type { SignInSubmitForm } from '../../model/sign-in-submit-form.model'
-import { subscribeToValueChangesOnForms } from 'src/app/shared/utils/subscribe-to-value-changes-on-forms.utils'
-import { hasNoSpaces } from 'src/app/shared/validators/has-no-spaces.validation'
+import type { SignInSubmitForm } from '../../models/sign-in-submit-form.model'
+import { subscribeToValueChangesOnForms } from 'src/app/shared/utils/subscribe-to-value-changes-on-forms.util'
+import { hasNoSpaces } from 'src/app/shared/validators/has-no-spaces.validator'
 import { hasOneNumber } from 'src/app/shared/validators/has-one-number.validator'
 
 @Component({
@@ -36,7 +36,7 @@ export class SignInFormComponent implements OnInit, OnDestroy {
     private authFacade: AuthFacade,
   ) {}
 
-  ngOnInit(): void {
+  public ngOnInit(): void {
     this.arraySubscriptions = subscribeToValueChangesOnForms([
       this.signInForm.controls.username,
       this.signInForm.controls.password,
