+++ src/app/auth/components/sign-up-form/sign-up-form.component.ts
@@ -4,14 +4,14 @@ import { TuiDay } from '@taiga-ui/cdk'
 import { type Subscription } from 'rxjs'
 
 import { AuthFacade } from '../../auth-store/auth.facade'
-import { transformRegistrationSubmitForm } from '../../utils/transform-registration-submit-form'
-import { toggleEnableStatusFields } from 'src/app/auth/dictionary/toggle-enable-status-fields.dictionary'
-import { Country } from 'src/app/shared/enum/country.enum'
-import { FormFields } from 'src/app/shared/enum/form-value.enum'
-import { subscribeToValueChangesOnForms } from 'src/app/shared/utils/subscribe-to-value-changes-on-forms.utils'
+import { transformRegistrationSubmitForm } from '../../utils/transform-registration-submit-form.util'
+import { toggleEnableStatusFields } from 'src/app/auth/utils/toggle-enable-status-fields.util'
+import { Country } from 'src/app/shared/enums/country.enum'
+import { FormFields } from 'src/app/shared/enums/form-value.enum'
+import { subscribeToValueChangesOnForms } from 'src/app/shared/utils/subscribe-to-value-changes-on-forms.util'
 import { birthValidator } from 'src/app/shared/validators/birth.validator'
 import { emailValidator } from 'src/app/shared/validators/email.validator'
-import { hasNoSpaces } from 'src/app/shared/validators/has-no-spaces.validation'
+import { hasNoSpaces } from 'src/app/shared/validators/has-no-spaces.validator'
 import { hasOneCharacter } from 'src/app/shared/validators/has-one-character.validator'
 import { hasOneLowerCaseCharacter } from 'src/app/shared/validators/has-one-lowercase-character.validator'
 import { hasOneNumber } from 'src/app/shared/validators/has-one-number.validator'
