+++ src/app/auth/utils/auth-alert.util.ts
@@ -1,7 +1,7 @@
 import type { TuiAlertService } from '@taiga-ui/core'
 import type { Subscription } from 'rxjs/internal/Subscription'
 
-import { AlertsStatus } from 'src/app/shared/enum/alets-status.enum'
+import { AlertsStatus } from 'src/app/shared/enums/alerts-status.enum'
 
 export const alertsAuth: Record<string, (alertService: TuiAlertService, action: string) => Subscription> = {
   true: (alertService: TuiAlertService, action: string): Subscription => {
