+++ src/app/shared/enums/store-feature-names.enum.ts
