+++ src/app/shared/validators/email.validator.ts
@@ -1,6 +1,6 @@
 import type { AbstractControl, ValidationErrors } from '@angular/forms'
 
-import { emailRegExp } from '../constants/regExp'
+import { emailRegExp } from '../constants/regExp.constant'
 
 export const emailValidator = (control: AbstractControl): ValidationErrors | null => {
   const isValid = emailRegExp.test(control.value as string)
