+++ src/app/shared/validators/has-no-spaces.validator.ts
@@ -1,0 +1,7 @@
+import type { AbstractControl, ValidationErrors } from '@angular/forms'
+
+export function hasNoSpaces(control: AbstractControl): ValidationErrors | null {
+  const isValid = !((control.value as string).startsWith(' ') || (control.value as string).endsWith(' '))
+
+  return isValid ? null : { error: 'The field cannot start and end with a space.' }
+}
