+++ src/app/shared/validators/has-one-lowercase-character.validator.ts
@@ -1,6 +1,6 @@
 import type { AbstractControl, ValidationErrors } from '@angular/forms'
 
-import { lowerCharacterRegExp } from '../constants/regExp'
+import { lowerCharacterRegExp } from '../constants/regExp.constant'
 
 export const hasOneLowerCaseCharacter = (control: AbstractControl): ValidationErrors | null => {
   const isValid = lowerCharacterRegExp.test(control.value as string)
