+++ src/app/shared/validators/has-one-number.validator.ts
@@ -1,6 +1,6 @@
 import type { AbstractControl, ValidationErrors } from '@angular/forms'
 
-import { numberRegExp } from '../constants/regExp'
+import { numberRegExp } from '../constants/regExp.constant'
 
 export function hasOneNumber(control: AbstractControl): ValidationErrors | null {
   const isValid = numberRegExp.test(control.value as string)
