+++ src/app/shared/validators/name.validator.ts
@@ -1,6 +1,6 @@
 import type { AbstractControl, ValidationErrors } from '@angular/forms'
 
-import { nameRegExp } from '../constants/regExp'
+import { nameRegExp } from '../constants/regExp.constant'
 
 export const nameValidator = (control: AbstractControl): ValidationErrors | null => {
   const isValid = nameRegExp.test(control.value as string)
